+++ ChangeLog.md
@@ -1,5 +1,11 @@
 # Revision history for zeolite-lang
 
+## 0.11.0.0  -- ????-??-??
+
+* **[dependency]** Switches from `parsec` to `megaparsec` for parsing, to allow
+  for better structuring of error messages. This might prevent installation when
+  it previously would have been possible.
+
 ## 0.10.0.0  -- 2020-12-05
 
 ### Language